requirements.txt

Django==3.0.5
django-crispy-forms==1.9.0
Pillow==7.1.1
requests==2.23.0
psycopg2>=2.7,<3.0